+++ src/Presentation/SmartStore.Web/Administration/Models/Customers/CustomerAddressModel.cs
@@ -6,6 +6,7 @@ namespace SmartStore.Admin.Models.Customers
     public class CustomerAddressModel : ModelBase
     {
         public int CustomerId { get; set; }
+        public string Username { get; set; }
 
         public AddressModel Address { get; set; }
     }