+++ changelog.md
@@ -16,6 +16,7 @@
 ### Improvements
 * #1663 Make MeasureDimension and MeasureWeight localizable.
 * #1600 Show an example currency value for custom formatting value changes.
+* Card desk instead of a grid for customer addresses on customer edit page.
 
 ### Bugfixes
 * Export: Fixed KeyNotFoundException when batch size was 1.